feat: engenheiro falado, economia bottom-up e acervo de voz em Opus
Reune o trabalho em andamento de varias frentes, e converte o acervo de voz
para Opus para que ele caiba no repositorio.

- Engenheiro falado (src-tauri/src/engenheiro/, commands/engenheiro/): rádio de
  equipe com push-to-talk. Dois caminhos por pergunta — o renderizador monta a
  resposta de pecas gravadas quando a forma e fixa (~0,7 s, de graca), e so cai
  no Gemini + TTS quando nao renderiza (~2,9 s). Cobre quebra, ritmo, tempo de
  volta, volta de referencia, classificacao e campeonato, com memoria, momento
  (o portao que cala o radio em largada, duelo e ultima volta) e tratamento.
- Acervo de voz: 3.943 pecas do engenheiro e 51 do spotter, geradas por
  engenheiro-pack.mjs/spotter-pack.mjs com a Chirp 3 HD e a cadeia de radio
  gravada dentro. O catalogo sai do Rust, entao o que o app pede e o que o
  gerador grava vem da mesma lista.
- Economia bottom-up (src-tauri/src/economia/): modulo novo ao lado de
  finance/, ainda NAO integrado — nasce isolado para ser medido antes de
  substituir. A conta parte de quilometros, litros, jogos de pneu e pessoas em
  vez de fracionar um orcamento por categoria escrito a mao.
- Telas v2: ficha de piloto, pre-temporada, minha equipe e a aba propria de
  myteam.
- Overlay e iRacing: canal de radio no overlay, race_monitor com tentativas e
  captura de corrida, spotter de frentes.

Acervo em Opus, e nao em WAV:

As 3.994 pecas somam 328 MB em PCM 24 kHz — num repositorio de 214 MB, e
crescendo a cada fala nova. WAV nao delta-comprime: cada regravacao somaria o
arquivo inteiro ao historico para sempre. Em Opus 32 kbps sao 29 MB (11,2x), e
32 kbps e folgado porque a cadeia de radio ja descartou antes o que o codec
teria a descartar. A emenda nao estala: o app decodifica cada peca em
AudioBuffer e toca cada uma no seu proprio BufferSource, com pausa deliberada
de 160 ms entre frases — nada e emendado sample a sample em execucao.

Os WAV seguem sendo os masters, no disco e fora do git. As ferramentas que
medem FORMA DE ONDA (engenheiro-auditar.mjs e as audicoes em scripts/tts-poc/)
continuam lendo o PCM, porque pico e silencio interno nao se medem na saida do
codec. Os guards de disco passaram a conferir o .opus: conferir o WAV passaria
na maquina de quem gerou o acervo e falharia em todo clone limpo.

Dois defeitos achados no caminho:

- vite.config.js exclui o acervo do inline. Em WAV a peca mais curta tinha
  ~24 kB e o teto de 4 kB do Vite nunca era atingido; em Opus, 432 pecas caem
  abaixo dele e viravam 1,68 MB de base64 dentro do chunk principal, parseado
  no startup e contornando a fila de carga preguicosa. O chunk caiu de 4,16 MB
  para 2,31 MB.
- O guard ptt-microfone-escolhido acusava prosa: o regex de `armar(` casava a
  palavra dentro de um comentario em portugues, num arquivo cuja chamada real
  sempre passou o deviceId. Os comentarios saem antes da varredura.

+++ src-tauri/src/constants/categories.rs
@@ -17,7 +17,10 @@ pub struct CategoryConfig {
     pub pilotos_por_equipe: u8,
     pub grid_total: u8,
     pub corridas_por_temporada: u8,
-    pub duracao_corrida_min: u8,
+    /// Duração nominal da prova (min). `u16` porque o Endurance chega a 360 — não cabe em `u8`.
+    /// Vale 0 no Endurance: lá a duração é sorteada por etapa em `resolve_race_duration` e a
+    /// fonte de verdade passa a ser `CalendarEntry::duracao_corrida_min`.
+    pub duracao_corrida_min: u16,
     pub monomarca: bool,
     pub multi_classe: bool,
     pub licenca_necessaria: Option<u8>,
